Rename plan locations reducer
We are changing to use a better, more descriptive name.
